+++ src/Source.h
@@ -47,6 +47,57 @@ public:
     return cur;
   }
 
+  static const char* skipBom(const char* begin, const char* end) {
+
+    /* Unicode Byte Order Marks
+       https://en.wikipedia.org/wiki/Byte_order_mark#Representations_of_byte_order_marks_by_encoding
+
+       00 00 FE FF: UTF-32BE
+       FF FE 00 00: UTF-32LE
+       FE FF:       UTF-16BE
+       FF FE:       UTF-16LE
+       EF BB BF:    UTF-8
+   */
+
+    switch(begin[0]) {
+      // UTF-32BE
+      case '\x00':
+        if (end - begin >= 4 && begin[1] == '\x00' && begin[2] == '\xFE' && begin[3] == '\xFF') {
+          return begin + 4;
+        }
+        break;
+
+      // UTF-8
+      case '\xEF':
+        if (end - begin >= 3 && begin[1] == '\xBB' && begin[2] == '\xBF') {
+          return begin + 3;
+        }
+        break;
+
+      // UTF-16BE
+      case '\xfe':
+        if (end - begin >= 2 && begin[1] == '\xff') {
+          return begin + 2;
+        }
+        break;
+
+      case '\xff':
+        if (end - begin >= 2 && begin[1] == '\xfe') {
+
+          // UTF-32 LE
+          if (end - begin >= 4 && begin[2] == '\x00' && begin[3] == '\x00') {
+            return begin + 4;
+          }
+
+          // UTF-16 LE
+          return begin + 2;
+        }
+        break;
+    }
+    return begin;
+  }
+
+
   static SourcePtr create(Rcpp::List spec);
 
 private:
@@ -59,5 +110,3 @@ private:
 };
 
 #endif
-
-